Revert "fix: depth proto"
This reverts commit 3be2b6e666648620a2f9b137820160a083f8e163.

+++ app/src/main/java/protobuf/magic/struct/VarintResult.java
@@ -1,3 +1,5 @@
 package protobuf.magic.struct;
 
-public record VarintResult(Long value, int length) {}
+import java.math.BigInteger;
+
+public record VarintResult(BigInteger value, int length) {}
